Feature/core functions

+++ app/Http/Controllers/Api/ProjectController.php
@@ -7,17 +7,28 @@ use App\Http\Requests\StoreProjectRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Project;
+use App\Models\Task;
 use Exception;
 use Auth;
 use DB;
 
 class ProjectController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
         try {
+
+            $search = $request->q ?? null;
+            $pageIndex = $request->pageIndex ?? 0;
+            $pageSize = $request->pageSize ?? 3;
+            $sortBy = $request->sortBy ?? 'name';
+            $sortDirection = $request->sortDirection ?? 'ASC';
             
-            $projects = Project::all();
+            $projects = Project::where( 'name', 'LIKE', '%' . $search . '%' )
+                            ->skip($pageIndex*$pageSize)
+                            ->take($pageSize)
+                            ->orderBy($sortBy, $sortDirection)
+                            ->get();
 
             return response()->json([
                 'status' => 200,
@@ -40,6 +51,13 @@ class ProjectController extends Controller
         DB::beginTransaction();
 
         try {
+
+            if(getRole() != 'PRODUCT_OWNER'){
+                return response()->json([
+                    'status' => 401,
+                    'message' => 'you do not have permission to access this API',
+                ], 401);
+            }
             
             $project = Project::create([
                 'name' => $request->name,
@@ -50,6 +68,7 @@ class ProjectController extends Controller
             return response()->json([
                 'status'    => 200,
                 'message'   => 'Project Successfully Created',
+                'data'      => $project->toArray()
             ], 200);
 
         } catch (Exception $e) {
@@ -90,6 +109,13 @@ class ProjectController extends Controller
 
         try {
 
+            if(getRole() != 'PRODUCT_OWNER'){
+                return response()->json([
+                    'status' => 401,
+                    'message' => 'you do not have permission to access this API',
+                ], 401);
+            }
+
             $project->name = $request->name;
             $project->save();
             
@@ -117,6 +143,21 @@ class ProjectController extends Controller
         DB::beginTransaction();
 
         try {
+
+            if(getRole() != 'PRODUCT_OWNER'){
+                return response()->json([
+                    'status' => 401,
+                    'message' => 'you do not have permission to access this API',
+                ], 401);
+            }
+
+            $tasks = Task::where('project_id', $project->id)->get();
+
+            if(!empty($tasks->toArray())){
+                foreach($tasks as $task){
+                    $task->delete();
+                }  
+            }
 
             $project->delete();
